refactor: Remove the use of Vectors and Slices for operands, instead pass the parameters directly to the function.
This eliminates the need for the majority of the memory allocations with each instruction.

refactor: Overhauled the RGAL instruction set, removed a lot of register specific commands like PUSHX, POPX, INCX, INCY etc.
Simplified some names and made them more consistent.
Changed some opcode operands to only work with registers/immediate for clarity and to compensate for the removal of some opcodes.

refactor: Split the majority of the code into separate modules for easier work in the future.

refactor: Updated the parser so that it extracts the exact operands for each opcode and constructs the completed enum variant.